refactor(StonePositionHandler): MainActiviy로부터 쿼리 분리
