+++ app/src/main/java/woowacourse/omok/database/stoneposition/StonePositionDbHelper.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok.database
+package woowacourse.omok.database.stoneposition
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
